samples: knot: Add timeout for re-transmission
This patch adds the basic structure for managing pending KNoT operation
responses. KNoT binary protocol must be serialized, a second command
should not be issued if a response is pending. REG, AUTH and SCH
requires a response from the peer, if response is not received each
state should re-send the previous command or change to ERROR state.

Signed-off-by: Claudio Takahasi <ckt@cesar.org.br>