samples: knot: Resend data only if not success
Resend data on online state after polling only if not receiving success
response.
Modify proxy_read() and proxy_force_send() so proxy_read() could
receive a boolean parameter to flag if value should be resend if no
data response for this sensor is received.
Add proxy_confirm_sent() to flag that the data was successfully
received.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>