samples: knot: Checking timeout at each state
This patch removes the "resend" variable from sm_run() so each state
is responsible for handling timeout ocurrences. Considering that, it
only turns the timeout timer on if some response is expected.
By doing that, it is possible to keep the timer off even when staying
on the same state.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>