samples: knot: Resend DATA_MSG if timed out
This patch adds the resend feature to timed out data messages on online
state.

It also perimits that messages as SET_DATA and GET_DATA do not
be ignored when the system is waiting a response on online state.

To make sure no sensor deadlocks the polling, it will start from the
next available sensor. Looping until getting to the first sensor again.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>