core: ot_config: Notify on OpenThread role change
Notify if the OpenThread Device role has changed.

Zephyr defines OPENTHREAD_CONFIG_MAX_STATECHANGE_HANDLERS as 2 at
`subsys/net/lib/openthread/platform/openthread-core-zephyr-config.h`,
allowing the usage of 2 callbacks for state changes.
As the OpenThread subsystem uses one of these callbacks, the second
callback is free to use.

If the new device role is not defined as CHILD, a warning is raised.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>