core: Use Semaphores to indicate connection status
Use k_sem to indicate the connection status between the net and proto
threads.

This update is necessary so the net thread can inform the proto thread
if the device is connected or not.

As alerts are no longer supported, a semaphore is suitable for this
function because its value can be read by proto without the necessity
to change it, differing from a mutex per example.

If the connection status is changed, the state machine will be started
or stopped by proto thread.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>