setup: cmakelists: Add Kconfig core file to Setup
Add core Kconfig file to Setup App a value for CONFIG_KNOT_LOG_LEVEL
will be defined.

This change is necessary as activating CONFIG_LOG for Setup App would
break compiling as CONFIG_KNOT_LOG_LEVEL was not defined.
The printed error is:

"'CONFIG_KNOT_LOG_LEVEL' undeclared here (not in a function) \
 LOG_MODULE_REGISTER(ot_config, CONFIG_KNOT_LOG_LEVEL);"

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>