core: net: Init UDP handler if using UDP
Add init, start and stop functions to control IPV6 over UDP in case of
flag CONFIG_UDP is set.

The peer's ipv6 is still required for UDP, so it is loaded from storage
at UDP handler init.

In case of CONFIG_UDP set, the TCP handler will be disabled.

Not sending messages or handling received messages.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>