core: storage: Merge storage.nrf52840
Merge storage.nrf52840 file to storage.c.

As KNoT is intent to run on multiple types of devices and with the same
behavior for all of these devices, there is no point in defining a file
for each board.

As there is a different behavior when running on QEMU, the file
storage.qemu was kept.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>