boards: Add support to nrf52840_pca10059 (dongle)
Add dts and conf overlay files to support the usage nrf52840_pca10059.

The dts files define the stock and debug flash configuration. While the
conf file sets the basic configuration for KNoT apps to run in the
current settings.

The original flash configuration is respected, being the sub-slots just
parts of the original slots.

As the dongle doesn't start the USB stack automatically, the USB
functions will be disabled (including logging).

Depending on the MCUBOOT version, it's possible to use MCUBOOT's DFU
as well.

The cli.py script was not updated so its necessary to manually flash
the generated signed.hex file in case of using a programmer that is
not supported by nrfjprog.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>