core: boards: Set NET RX stack size by board
The NET TX and RX stack sizes should vary depending on the board and
its capabilities. This should be made based on the board to be used.
The overlay-knot-ot.conf file defines the minimum stack size for OT
while the maximum size is board based.

Sets CONFIG_NET_RX_STACK_SIZE for the boards nrf52840_pca10056 (DK) and
nrf52840_pca10059 (Dongle).

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>