scripts: cli: Flash image after build as an option
Flashing image through 'make flash' passed the wrong idea that it would
have an different nature than 'make', when it would actually just build
the image as usually but flash it in the end.

Set the 'flash after build' behavior as an option instead of a command
to fix this distortion.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>