scripts: cli: Add DFU flash support to dongle
Set default flash mode as through Open DFU Bootloader for dongle and
permit the user to set the target board where the DFU driver is
connected to.

Add executable nrfutil tool for linux generated with PyInstaller:
Compiling info:
 nrfutil version: 65bcd5c0038083b24c770fc0ddfcf17f566f76cf
 nrfutil repository:
    https://github.com/NordicSemiconductor/pc-nrfutil
 Python version: 2.7.12
 PyInstaller version: 3.4
 Compilation date: 2019-05-21

Create alias variables for board names so they can be referred as
constants.

Refer to MCUBOOT stock files and Board IDs based on board alias.

Modify KnotSDK.__flash() so it requires a device port argumment to be
passed. Abort if the target board is the dongle and no port was passed.

Use nrfjprog if using DK and nrfutil if using Dongle.

Always flash full project if using Dongle. This is necessary as the
Open DFU Bootloader clears the MCUBOOT when updating the firmware. So
the new firmware should also contain MCUBOOT.

Add '--port' option to 'make', 'mcuboot' and 'flash' commands.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>