scripts: cli: Split 'set' commands group
Split 'set' commands group into 'board' and 'ot-path' commands so it
won't be necessary for the user to type 'set' to define a default value.

Add --list flag to 'board' command that prints supported boards as the
missing argument error needs to be handled by the command, not the
group anymore.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>