core: nrf52840_pca10059: Disable UART interruption
Disable UART interruption for nrf52840_pca10059 to avoid conflict with
OpenThread.

This requires the USB stack to be disabled too.

This is necessary as there is a significant reduction in connection
range if it is on.

A possible workaround is to add a "USB-overlay" overlay file that can be
added to the project so USB can be used if the user wants to.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>