boards: Add support to KNoT Mesh
Add support to knot_mesh board based on Dongle.

Using the same mcuboot image as nrf52840_pca10059.
This has to be fixed so the push button 0 will be assigned to the
correct pin.

It assumes that a similar DFU bootloader from Dongle's is used.
Although the same bootloader is not suggested as the LED 0 is assigned
to a different pin.

Signed-off-by: Joao Cordeiro <jvcc@cesar.org.br>