scripts: cli: Check for a local DTS overlay file
Check for a DTS overlay file in the application folder whenever the user
does not provide an external path using the option '--dts_ovl_path' on
'make' command.

The overlay file should be named as <BOARD>.overlay and it should
be placed on the root application folder.

If the file does not exist, the system will use the default DTS files.

Signed-off-by: Allysson Lukas <alla@cesar.org.br>