scripts: cli: Check for an alternative DTS overlay
Check for a DTS overlay file named as '<BOARD_ALIASES>.overlay' in the
main application directory whenever the user does not provide an
external path using the option '--dts_ovl_path' on 'make' command and
the '<BOARD>.overlay' file is not found.

If the file does not exist, the system will use the default DTS files.

Signed-off-by: Allysson Lukas <alla@cesar.org.br>